Extract `WrapBoundBox3d` to a distinct file
* Create a new file `wrap_shape3d.cpp` to house wrapping code for 3-D geometry
  entity.
* Move the C++ wrapper class `WrapBoundBox3d` from file `wrap_World.cpp` to
  `wrap_shapd3d.cpp`.

Claude code command:
* Extract `WrapBoundBox3d` from `wrap_World.cpp` to a new file
  `wrap_shape3d.cpp`
